requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asyncio==3.4.3
bidict==0.22.1
BTrees==5.2
certifi==2022.12.7
cffi==1.16.0
chardet==4.0.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==2.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
neo-api-client @ git+https://github.com/Kotak-Neo/kotak-neo-api.git@67143c58f29da9572cdbb273199852682a0019d5
numpy==1.24.2
orjson==3.10.3
packaging==24.0
pandas==2.0.0
pluggy==1.5.0
pycparser==2.22
pydantic==2.7.1
pydantic-extra-types==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.18.0
pyjsparser==2.7.1
PyJWT==2.6.0
pytest==8.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.25.1
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
transaction==4.0
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
urllib3==1.26.14
uvicorn==0.29.0
watchfiles==0.21.0
websocket-client==1.5.1
websockets==12.0
zc.lockfile==3.0.post1
ZConfig==4.1
zodbpickle==3.3
zope.interface==6.3
